mcp/app/config.py
import os
from pydantic_settings import BaseSettings
from dotenv import load_dotenv


class Settings(BaseSettings):
    load_dotenv()

    # AI
    openai_api_key: str = os.getenv("OPENAI_API_KEY", "")
    OPENAI_EMBEDDING_MODEL: str = "text-embedding-3-large"
    OPENAI_MODEL: str = "gpt-4o-mini"

settings = Settings()